refactor: move package script to JsiiProject
