+++ .gitattributes
@@ -1,7 +1,7 @@
 # ~~ Generated by projen. To modify, edit .projenrc.ts and run "node ./projen.js".
 
 *.snap linguist-generated
-/./scripts/package.js linguist-generated
+/./projenrc/package.js linguist-generated
 /.devcontainer.json linguist-generated
 /.eslintrc.json linguist-generated
 /.gitattributes linguist-generated
