+++ .projen/files.json
@@ -24,7 +24,7 @@
     ".vscode/launch.json",
     "LICENSE",
     "projen.js",
-    "scripts/package.js",
+    "projenrc/package.js",
     "tsconfig.dev.json"
   ],
   "//": "~~ Generated by projen. To modify, edit .projenrc.ts and run \"node ./projen.js\"."
